rename: ui.lua → appearance.lua
